requirements/base.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
django-filter==2.2.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
drf-yasg==1.17.1
swagger-spec-validator==2.5.0
jupyterlab==2.0.1
django-extensions==2.2.8
sentry-sdk==0.14.2
pytest-django==3.8.0
mock==4.0.1
pandas==1.0.1
beautifulsoup4==4.8.2
model-bakery==1.1.0
gekko==0.2.6
selenium==3.141.0
yfinance==0.1.54
awscli==1.18.26
django-cors-headers==3.2.1
matplotlib==3.2.1
celery==4.4.2
redis==3.4.1
django-celery-beat==2.0.0
django-redis==4.11.0
CurrencyConverter==0.14.1
django-environ==0.4.5